requirements.txt
Flask==2.0.1
requests==2.26.0
matplotlib==3.4.3
python-dotenv==0.19.0
blinker==1.9.0
click==8.1.7
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
numpy==1.23.5  # Updated to a compatible version for Python 3.11
packaging==24.2
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
six==1.16.0
Werkzeug==2.0.2
